More precise makefile dependency rule
The rule for validating examples had a dependency on ALL frontend
schemas. With this rule, whenever any schema changes (even one for
another format), all examples would be re-validated.

This changes the dependency to be more precise, so that build rule for
the the `.valid` document which records the validation outcome depends
on the specific example being validated and the specific frontend schema
for that format.

I also change the argument to the validate command to specify the schema
path explicitly using the `--schema` option. I think that previously I
hadn't known how to calculate the correct path to the schema in the
makefile, so I had relied on the ruby script to work it out itself.