Add content schema type for Service Manual Section documents
Service Manual Section is a format that will be published via a rake task in
the service-manual-publisher initially, until we build a UI for it.
It represents a page that groups other pages in a way similar to Topic pages.
We're building a custom format as we use V2 publishing APIs and the generic
topics are not quite ready to be used in an ideal way.

The latest prototype of a Service Manual Section:
http://sm-11.herokuapp.com/agile-delivery/