Add history mode fields to World Location News
This is WIP, storing `is_historic` is a flat value in the details
hash will cause problems when a government changes - as documents
belonging to the (now) former government won't immediately be marked
as historic without a republished.

One solution is referencing the government through the `links` hash
so that it is expanded at fetch-time from content store, and the
frontend can use the government response (including end date) to work
out if a document is historic.