Derive is_historic from government
Previously we had a `is_historic` field, which would be computed at
publication time. If the government changed after this the documents
belonging to that government would need to be republished.

Instead we provide a government object, which is marked as current or
not, which a frontend can use to derive if a document should be
considered historic.

Including the government in the details hash still requires the document
to be republished - but it's planned to move government into the links
hash - which is expanded in the content-store at read time, which will
avoid the need to republish. This will require governments to be in
the content store and possibly have a URL.

In the mean time we have a government object that is similar to what we'd
expect back from a link-expanded government.

This also adds a slug field for government, which is used for analytics.