Add a summary to service_manual_guide
We are extending service_manual_guide to also handle the "points page" in the service manual. A points page represents one of the points in the service standard. The design of the points page requires a lede just after the title. Because these lede / standfirst sections are fairly common across GOV.UK we think it is acceptable to give this to the other two pages that use this format, the guide page and the guide community page.